Collapsed fixtures dir from tree view
