+++ packages/react-cosmos-playground2/src/plugins/Nav/FixtureTree/shared.js
@@ -1,0 +1,8 @@
+// @flow
+
+export type TreeNode<T> = {
+  values?: T[],
+  children: {
+    [dir: string]: TreeNode<T>
+  }
+};
